fix import, create defi-didDomain
