+++ lib/blockchain/util.ts
@@ -14,7 +14,6 @@ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.*/
 
 import hash from 'hash.js'
-import { TransitionParams, TransitionValue } from './tyronzil'
 import * as zutil from '@zilliqa-js/util'
 
 export default class Util {
@@ -41,98 +40,6 @@ export default class Util {
             h_ = h_ + h
         }
         return [hash_, h_]
-    }
-
-    public static async AddLiquidity(
-        signature: any,
-        id: string,
-        amount: string,
-        tyron: TransitionValue
-    ) {
-        const params = Array()
-
-        const sig: TransitionParams = {
-            vname: 'signature',
-            type: 'Option ByStr64',
-            value: signature,
-        }
-        params.push(sig)
-
-        const id_: TransitionParams = {
-            vname: 'addrID',
-            type: 'String',
-            value: id,
-        }
-        params.push(id_)
-
-        const amount_: TransitionParams = {
-            vname: 'amount',
-            type: 'Uint128',
-            value: amount,
-        }
-        params.push(amount_)
-
-        const tyron_: TransitionParams = {
-            vname: 'tyron',
-            type: 'Option Uint128',
-            value: tyron,
-        }
-        params.push(tyron_)
-        return params
-    }
-
-    public static async RemoveLiquidity(
-        signature: any,
-        id: string,
-        amount: string,
-        minZil: string,
-        minToken: string,
-        tyron: TransitionValue
-    ) {
-        const params = Array()
-
-        const sig: TransitionParams = {
-            vname: 'signature',
-            type: 'Option ByStr64',
-            value: signature,
-        }
-        params.push(sig)
-
-        const id_: TransitionParams = {
-            vname: 'addrID',
-            type: 'String',
-            value: id,
-        }
-        params.push(id_)
-
-        const amount_: TransitionParams = {
-            vname: 'amount',
-            type: 'Uint128',
-            value: amount,
-        }
-        params.push(amount_)
-
-        const minzil: TransitionParams = {
-            vname: 'minZilAmount',
-            type: 'Uint128',
-            value: minZil,
-        }
-        params.push(minzil)
-
-        const mintoken: TransitionParams = {
-            vname: 'minTokenAmount',
-            type: 'Uint128',
-            value: minToken,
-        }
-        params.push(mintoken)
-
-        const tyron_: TransitionParams = {
-            vname: 'tyron',
-            type: 'Option Uint128',
-            value: tyron,
-        }
-        params.push(tyron_)
-        return params
     }
 
     public static async HashString(s: string) {
